# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Unisphere API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getCurrentUser
      tags: [users]
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
    patch:
      operationId: updateCurrentUser
      tags: [users]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateUserRequest" }
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
...
  /users/{id}:
    get:
      operationId: getUserById
      tags: [users]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: User
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
  /users/{id}/follow:
    post:
      operationId: followUser
      tags: [users]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OkResponse" }
    delete:
      operationId: unfollowUser
      tags: [users]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OkResponse" }
  /posts:
    get:
      operationId: getPosts
      tags: [posts]
      responses:
        "200":
          description: Posts feed
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Post" }
    post:
      operationId: createPost
      tags: [posts]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreatePostRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Post" }
  /posts/{id}:
    delete:
      operationId: deletePost
      tags: [posts]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OkResponse" }
...
  /posts/{id}/like:
    post:
      operationId: likePost
      tags: [posts]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Post" }
    delete:
      operationId: unlikePost
      tags: [posts]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Post" }
  /posts/{id}/comments:
    post:
      operationId: createComment
      tags: [posts]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateCommentRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Comment" }
...
  /societies/{id}/join:
    post:
      operationId: joinSociety
      tags: [societies]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Society" }
    delete:
      operationId: leaveSociety
      tags: [societies]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Society" }
...
  /storage/uploads/request-url:
    post:
      tags: [Storage]
      operationId: requestUploadUrl
      summary: Request a presigned URL for file upload
      description: |
        Returns a presigned GCS URL for direct upload. The client sends JSON
        metadata here, then uploads the file directly to the returned URL.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadUrlRequest'
      responses:
        '200':
          description: Presigned upload URL generated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadUrlResponse'
        '400':
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Failed to generate upload URL.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
components:
  schemas:
...
    OkResponse:
      type: object
      required: [ok]
      properties:
        ok: { type: boolean }
    User:
      type: object
      required: [id, username, name, email, major, avatar, cover, bio, tags, followersCount, followingCount, postsCount]
      properties:
        id: { type: integer }
        username: { type: string }
        name: { type: string }
        email: { type: string }
        major: { type: string }
        avatar: { type: string }
        cover: { type: string }
        bio: { type: string }
        tags:
          type: array
          items: { type: string }
        followersCount: { type: integer }
        followingCount: { type: integer }
        postsCount: { type: integer }
        createdAt: { type: string }
    UpdateUserRequest:
      type: object
      properties:
        name: { type: string }
        major: { type: string }
        avatar: { type: string }
        cover: { type: string }
        bio: { type: string }
        tags:
          type: array
          items: { type: string }
...
    Comment:
      type: object
      required: [id, postId, author, avatar, text, time]
      properties:
        id: { type: integer }
        postId: { type: integer }
        author: { type: string }
        avatar: { type: string }
        text: { type: string }
        time: { type: string }
    Post:
      type: object
      required: [id, author, major, avatar, content, hashtags, likes, comments, isLiked, isOwn, commentList]
      properties:
        id: { type: integer }
        author: { type: string }
        major: { type: string }
        avatar: { type: string }
        content: { type: string }
        hashtags: { type: string }
        image: { type: string, nullable: true }
        likes: { type: integer }
        comments: { type: integer }
        isLiked: { type: boolean }
        isOwn: { type: boolean }
        commentList:
          type: array
          items: { $ref: "#/components/schemas/Comment" }
    CreatePostRequest:
      type: object
      required: [content]
      properties:
        content: { type: string }
        hashtags: { type: string }
        image: { type: string, nullable: true }
    CreateCommentRequest:
      type: object
      required: [text]
      properties:
        text: { type: string }
    Society:
      type: object
      required: [id, name, desc, icon, members, joined]
      properties:
        id: { type: integer }
        name: { type: string }
        desc: { type: string }
        icon: { type: string }
        members: { type: integer }
        joined: { type: boolean }
...
    UploadUrlRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name:
          type: string
          minLength: 1
          description: Original file name.
        size:
          type: integer
          minimum: 1
          description: File size in bytes.
        contentType:
          type: string
          minLength: 1
          description: MIME type of the file (e.g. `image/jpeg`).
    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL:
          type: string
          format: uri
          description: Presigned GCS URL for PUT upload.
        objectPath:
          type: string
          description: Normalized object path (e.g. `/objects/uploads/uuid`). Store this in your database.
        metadata:
          $ref: '#/components/schemas/UploadUrlRequest'
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string