# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Unisphere API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  username: string;
  name: string;
  email: string;
  major: string;
  avatar: string;
  cover: string;
  bio: string;
  tags: string[];
  followersCount: number;
  followingCount: number;
  postsCount: number;
  createdAt?: string;
}